requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
bcrypt==3.1.7
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.113.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
iniconfig==2.0.0
packaging==24.1
passlib==1.7.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
pytest==8.3.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.38.5
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
uvicorn==0.15.0
